requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
tflite-runtime==2.10.0  # Versi ringan untuk TFLite
numpy==1.19.5
Pillow==8.4.0  # Untuk PIL (Image processing)
